Move more code from MicroClimate class to MicroClimateZone class
